# Conflicts:
#	utils/filter_users.test.ts
#	utils/filter_users.ts

+++ components/admin_console/group_settings/group_row.jsx
@@ -11,7 +11,7 @@ import {localizeMessage} from 'utils/utils.jsx';
 import CheckboxCheckedIcon from 'components/widgets/icons/checkbox_checked_icon.jsx';
 import LoadingSpinner from 'components/widgets/loading/loading_spinner';
 
-export default class GroupRow extends React.Component {
+export default class GroupRow extends React.PureComponent {
     static propTypes = {
         primary_key: PropTypes.string.isRequired,
         name: PropTypes.string.isRequired,
@@ -20,6 +20,7 @@ export default class GroupRow extends React.Component {
         checked: PropTypes.bool,
         failed: PropTypes.bool,
         onCheckToggle: PropTypes.func,
+        readOnly: PropTypes.bool,
         actions: PropTypes.shape({
             link: PropTypes.func.isRequired,
             unlink: PropTypes.func.isRequired,
@@ -34,13 +35,18 @@ export default class GroupRow extends React.Component {
     }
 
     onRowClick = () => {
+        if (this.props.readOnly) {
+            return;
+        }
         this.props.onCheckToggle(this.props.primary_key);
     }
 
     linkHandler = async (e) => {
         e.stopPropagation();
         e.preventDefault();
-
+        if (this.props.readOnly) {
+            return;
+        }
         this.setState({loading: true});
         await this.props.actions.link(this.props.primary_key);
         this.setState({loading: false});
@@ -49,6 +55,9 @@ export default class GroupRow extends React.Component {
     unlinkHandler = async (e) => {
         e.stopPropagation();
         e.preventDefault();
+        if (this.props.readOnly) {
+            return;
+        }
         this.setState({loading: true});
         await this.props.actions.unlink(this.props.primary_key);
         this.setState({loading: false});
@@ -117,6 +126,7 @@ export default class GroupRow extends React.Component {
                 <a
                     href='#'
                     onClick={this.unlinkHandler}
+                    className={this.props.readOnly ? 'disabled' : ''}
                 >
                     <i className='icon fa fa-link'/>
                     <FormattedMessage
@@ -145,6 +155,7 @@ export default class GroupRow extends React.Component {
             <a
                 href='#'
                 onClick={this.linkHandler}
+                className={this.props.readOnly ? 'disabled' : ''}
             >
                 <i className='icon fa fa-unlink'/>
                 <FormattedMessage
